--- contracts/v4/Migrations.sol
@@ -1,24 +1,0 @@
-pragma solidity ^0.4.4;
-
-
-contract Migrations {
-    address public owner;
-    uint public last_completed_migration;
-
-    modifier restricted() {
-        if (msg.sender == owner) _;
-    }
-
-    function Migrations() public {
-        owner = msg.sender;
-    }
-
-    function setCompleted(uint completed) public restricted {
-        last_completed_migration = completed;
-    }
-
-    function upgrade(address new_address) public restricted {
-        Migrations upgraded = Migrations(new_address);
-        upgraded.setCompleted(last_completed_migration);
-    }
-}
