+++ helpers/migVals.js
@@ -37,9 +37,9 @@ var { gasToCash, _ } = require('./utils')
 
 async function updateCloversController({
   cloversController,
-  // curationMarket,
   clubTokenController,
-  simpleCloversMarket
+  simpleCloversMarket,
+  verbose
 }) {
 
   var totalGas = utils.toBN('0')
@@ -51,85 +51,83 @@ async function updateCloversController({
   // -w stakeAmount
   // -w stakePeriod
   // -w payMultiplier
-  // console.log(_ + 'cloversController.updateCurationMarket')
-  // var tx = await cloversController.updateCurationMarket(curationMarket.address)
 
   var currentClubTokenControllerAddress = await cloversController.clubTokenController()
   if (currentClubTokenControllerAddress.toLowerCase() !== clubTokenController.address.toLowerCase()) {
-    console.log(_ + `cloversController.updateClubTokenController from ${currentClubTokenControllerAddress} to ${clubTokenController.address}`)
+    verbose && console.log(_ + `cloversController.updateClubTokenController from ${currentClubTokenControllerAddress} to ${clubTokenController.address}`)
     var tx = await cloversController.updateClubTokenController(
       clubTokenController.address
     )
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'clubTokenController hasnt changed')
+    verbose && console.log(_ + 'clubTokenController hasnt changed')
   }
 
 
   var currentSimpleCloversMarket = await cloversController.simpleCloversMarket()
   if (currentSimpleCloversMarket.toLowerCase() !== simpleCloversMarket.address.toLowerCase()) {
-    console.log(_ + `cloversController.updateSimpleCloversMarket from ${currentSimpleCloversMarket} to ${simpleCloversMarket.address}`)
+    verbose && console.log(_ + `cloversController.updateSimpleCloversMarket from ${currentSimpleCloversMarket} to ${simpleCloversMarket.address}`)
     var tx = await cloversController.updateSimpleCloversMarket(
       simpleCloversMarket.address
     )
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'simpleCloversMarket hasnt changed')
+    verbose && console.log(_ + 'simpleCloversMarket hasnt changed')
   }
 
   var currentOracle = await cloversController.oracle()
   if (currentOracle.toLowerCase() !== vals.oracle.toLowerCase()) {
-    console.log(_ + `cloversController.updateOracle from ${currentOracle} to ${vals.oracle}`)
+    verbose && console.log(_ + `cloversController.updateOracle from ${currentOracle} to ${vals.oracle}`)
     var tx = await cloversController.updateOracle(vals.oracle)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'oracle hasnt changed')
+    verbose && console.log(_ + 'oracle hasnt changed')
   }
 
   var currentPayMultiplier = await cloversController.payMultiplier()
   if (!currentPayMultiplier.eq(vals.payMultiplier)) {
-    console.log(_ + `cloversController.updatePayMultipier from ${currentPayMultiplier} to ${vals.payMultiplier}`)
+    verbose && console.log(_ + `cloversController.updatePayMultipier from ${currentPayMultiplier} to ${vals.payMultiplier}`)
     var tx = await cloversController.updatePayMultipier(vals.payMultiplier)  
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'payMultiplier hasnt changed')
+    verbose && console.log(_ + 'payMultiplier hasnt changed')
   }
 
 
   var currentPriceMultiplier = await cloversController.priceMultiplier()
   if (!currentPriceMultiplier.eq(vals.priceMultiplier)) {
-    console.log(_ + `cloversController.updatePriceMultipier from ${currentPriceMultiplier} to ${vals.priceMultiplier}`)
+    verbose && console.log(_ + `cloversController.updatePriceMultipier from ${currentPriceMultiplier} to ${vals.priceMultiplier}`)
     var tx = await cloversController.updatePriceMultipier(vals.priceMultiplier)  
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'priceMultiplier hasnt changed')
+    verbose && console.log(_ + 'priceMultiplier hasnt changed')
   }
 
 
   var currentBasePrice = await cloversController.basePrice()
   if (!currentBasePrice.eq(vals.basePrice)) {
-    console.log(_ + `cloversController.updateBasePrice from ${currentBasePrice} to ${vals.basePrice}`)
+    verbose && console.log(_ + `cloversController.updateBasePrice from ${currentBasePrice} to ${vals.basePrice}`)
     var tx = await cloversController.updateBasePrice(vals.basePrice)  
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'basePrice hasnt changed')
+    verbose && console.log(_ + 'basePrice hasnt changed')
   }
 
 
   var currentPaused = await cloversController.paused()
   if (currentPaused !== vals.paused) {
-    console.log(_ + `cloversController.updatePaused from ${currentPaused} to ${vals.paused}`)
+    verbose && console.log(_ + `cloversController.updatePaused from ${currentPaused} to ${vals.paused}`)
     var tx = await cloversController.updatePaused(vals.paused)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'paused hasnt changed')
+    verbose && console.log(_ + 'paused hasnt changed')
   }
 
   return totalGas
@@ -139,7 +137,8 @@ async function updateClubTokenController({
   clubTokenController,
   // curationMarket,
   simpleCloversMarket,
-  accounts
+  accounts,
+  verbose
 }) {
   var totalGas = utils.toBN('0')
 
@@ -154,62 +153,62 @@ async function updateClubTokenController({
 
   var currentSimpleCloversMarket = await clubTokenController.simpleCloversMarket()
   if (currentSimpleCloversMarket.toLowerCase() !== simpleCloversMarket.address.toLowerCase()) {
-    console.log(_ + `clubTokenController.updateSimpleCloversMarket from ${currentSimpleCloversMarket} to ${simpleCloversMarket.address}`)
+    verbose && console.log(_ + `clubTokenController.updateSimpleCloversMarket from ${currentSimpleCloversMarket} to ${simpleCloversMarket.address}`)
     var tx = await clubTokenController.updateSimpleCloversMarket(
       simpleCloversMarket.address
     )
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'simpleCloversMarket hasnt changed')
+    verbose && console.log(_ + 'simpleCloversMarket hasnt changed')
   }
 
-  // console.log(_ + 'clubTokenController.updateCurationMarket')
+  // verbose && console.log(_ + 'clubTokenController.updateCurationMarket')
   // var tx = await clubTokenController.updateCurationMarket(
   //   curationMarket.address
   // )
 
   var currentReserveRatio = await clubTokenController.reserveRatio()
   if (!currentReserveRatio.eq(vals.reserveRatio)) {
-    console.log(_ + `clubTokenController.updateReserveRatio from ${currentReserveRatio} to ${vals.reserveRatio}`)
+    verbose && console.log(_ + `clubTokenController.updateReserveRatio from ${currentReserveRatio} to ${vals.reserveRatio}`)
     var tx = await clubTokenController.updateReserveRatio(vals.reserveRatio)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'reserveRatio hasnt changed')
+    verbose && console.log(_ + 'reserveRatio hasnt changed')
   }
 
   var currentVirtualSupply = await clubTokenController.virtualSupply()
   if (!currentVirtualSupply.eq(vals.virtualSupply)) {
-    console.log(_ + `clubTokenController.updateVirtualSupply from ${currentVirtualSupply.toString()} to ${vals.virtualSupply.toString()}`)
+    verbose && console.log(_ + `clubTokenController.updateVirtualSupply from ${currentVirtualSupply.toString()} to ${vals.virtualSupply.toString()}`)
     var tx = await clubTokenController.updateVirtualSupply(vals.virtualSupply)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'virtualSupply hasnt changed')
+    verbose && console.log(_ + 'virtualSupply hasnt changed')
   }
 
   var currentVirtualBalance = await clubTokenController.virtualBalance()
   if (!currentVirtualBalance.eq(vals.virtualBalance)) {
-    console.log(_ + `clubTokenController.updateVirtualBalance from ${currentVirtualBalance.toString()} to ${vals.virtualBalance.toString()}`)
+    verbose && console.log(_ + `clubTokenController.updateVirtualBalance from ${currentVirtualBalance.toString()} to ${vals.virtualBalance.toString()}`)
     var tx = await clubTokenController.updateVirtualBalance(vals.virtualBalance)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'virtualBalance hasnt changed')
+    verbose && console.log(_ + 'virtualBalance hasnt changed')
   }
 
   var currentPaused = await clubTokenController.paused()
   if (currentPaused !== vals.paused) {
-    console.log(_ + `clubTokenController.updatePaused from ${currentPaused} to ${vals.paused}`)
+    verbose && console.log(_ + `clubTokenController.updatePaused from ${currentPaused} to ${vals.paused}`)
     var tx = await clubTokenController.updatePaused(vals.paused)
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + 'paused hasnt changed')
+    verbose && console.log(_ + 'paused hasnt changed')
   }
 
-  console.log(_ + 'remove as admins to clubTokenController')
+  verbose && console.log(_ + 'remove as admins to clubTokenController')
   // await addAsAdmin(clubTokenController, accounts)
   var gasUsed = await removeAsAdmin(clubTokenController, accounts)
   totalGas = totalGas.add(gasUsed)
@@ -219,101 +218,101 @@ async function updateClubTokenController({
 
 
 
-async function addAsAdmin(contract, accounts) {
+async function addAsAdmin(contract, accounts, verbose) {
 
   var secondOwner = await contract.isAdmin(accounts[0])
   if (!secondOwner) {
-    console.log(_ + `adding ${accounts[0]} as admin in contract`)
+    verbose && console.log(_ + `adding ${accounts[0]} as admin in contract`)
     await contract.transferAdminship(accounts[0])
   } else {
-    console.log(_ + `${accounts[0]} is already an admin in contract`)
+    verbose && console.log(_ + `${accounts[0]} is already an admin in contract`)
   }
 
 
   var secondOwner = await contract.isAdmin(accounts[1])
   if (!secondOwner) {
-    console.log(_ + `adding ${accounts[1]} as admin in contract`)
+    verbose && console.log(_ + `adding ${accounts[1]} as admin in contract`)
     await contract.transferAdminship(accounts[1])
   } else {
-    console.log(_ + `${accounts[1]} is already an admin in contract`)
+    verbose && console.log(_ + `${accounts[1]} is already an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[2])
   if (!secondOwner) {
-    console.log(_ + `adding ${accounts[2]} as admin in contract`)
+    verbose && console.log(_ + `adding ${accounts[2]} as admin in contract`)
     await contract.transferAdminship(accounts[2])
   } else {
-    console.log(_ + `${accounts[2]} is already an admin in contract`)
+    verbose && console.log(_ + `${accounts[2]} is already an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[3])
   if (!secondOwner) {
-    console.log(_ + `adding ${accounts[3]} as admin in contract`)
+    verbose && console.log(_ + `adding ${accounts[3]} as admin in contract`)
     await contract.transferAdminship(accounts[3])
   } else {
-    console.log(_ + `${accounts[3]} is already an admin in contract`)
+    verbose && console.log(_ + `${accounts[3]} is already an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[4])
   if (!secondOwner) {
-    console.log(_ + `adding ${accounts[4]} as admin in contract`)
+    verbose && console.log(_ + `adding ${accounts[4]} as admin in contract`)
     await contract.transferAdminship(accounts[4])
   } else {
-    console.log(_ + `${accounts[4]} is already an admin in contract`)
+    verbose && console.log(_ + `${accounts[4]} is already an admin in contract`)
   }
 }
 
-async function removeAsAdmin(contract, accounts) {
+async function removeAsAdmin(contract, accounts, verbose) {
   var totalGas = utils.toBN('0')
 
   var secondOwner = await contract.isAdmin(accounts[1])
   if (secondOwner) {
-    console.log(_ + `removing ${accounts[1]} as admin in contract`)
+    verbose && console.log(_ + `removing ${accounts[1]} as admin in contract`)
     var tx = await contract.renounceAdminship(accounts[1])
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + `${accounts[1]} is already removed as an admin in contract`)
+    verbose && console.log(_ + `${accounts[1]} is already removed as an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[2])
   if (secondOwner) {
-    console.log(_ + `removing ${accounts[2]} as admin in contract`)
+    verbose && console.log(_ + `removing ${accounts[2]} as admin in contract`)
     var tx = await contract.renounceAdminship(accounts[2])
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + `${accounts[2]} is already removed as an admin in contract`)
+    verbose && console.log(_ + `${accounts[2]} is already removed as an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[3])
   if (secondOwner) {
-    console.log(_ + `removing ${accounts[3]} as admin in contract`)
+    verbose && console.log(_ + `removing ${accounts[3]} as admin in contract`)
     var tx = await contract.renounceAdminship(accounts[3])
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + `${accounts[3]} is already removed as an admin in contract`)
+    verbose && console.log(_ + `${accounts[3]} is already removed as an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[4])
   if (secondOwner) {
-    console.log(_ + `removing ${accounts[4]} as admin in contract`)
+    verbose && console.log(_ + `removing ${accounts[4]} as admin in contract`)
     var tx = await contract.renounceAdminship(accounts[4])
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + `${accounts[4]} is already removed as an admin in contract`)
+    verbose && console.log(_ + `${accounts[4]} is already removed as an admin in contract`)
   }
 
   var secondOwner = await contract.isAdmin(accounts[0])
   if (secondOwner) {
-    console.log(_ + `removing ${accounts[0]} as admin in contract`)
+    verbose && console.log(_ + `removing ${accounts[0]} as admin in contract`)
     var tx = await contract.renounceAdminship(accounts[0])
-    gasToCash(tx.receipt.gasUsed)
+    verbose && gasToCash(tx.receipt.gasUsed)
     totalGas = totalGas.add(utils.toBN(tx.receipt.gasUsed))
   } else {
-    console.log(_ + `${accounts[0]} is already removed as an admin in contract`)
+    verbose && console.log(_ + `${accounts[0]} is already removed as an admin in contract`)
   }
   return totalGas
 }
